Rename taskmanagement to examplefeature
