+++ walking-skeleton/src/main/java/com/example/application/examplefeature/package-info.java
@@ -15,7 +15,7 @@
 //#if ui.framework == "hilla"
 @NonNullApi // Until https://github.com/vaadin/hilla/issues/2612 has been fixed
 //#endif
-package com.example.application.taskmanagement;
+package com.example.application.examplefeature;
 // TODO Remove this package once you have added real features
 
 import org.jspecify.annotations.NullMarked;
